% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_adduct_rules.R
\name{adduct_rules_neg}
\alias{adduct_rules_neg}
\title{Create list with all adduct calculation rules (negative ion model)

This function returns a list with rules for the calculation of adducts. 
It is required for the calculation of adduct m/z values from exact masses 
and the other way round. This list contains all rules for the negative 
ionization mode.}
\usage{
adduct_rules_neg()
}
\description{
Create list with all adduct calculation rules (negative ion model)

This function returns a list with rules for the calculation of adducts. 
It is required for the calculation of adduct m/z values from exact masses 
and the other way round. This list contains all rules for the negative 
ionization mode.
}
\seealso{
\code{\link{adduct_rules_all}}

\code{\link{adduct_rules_pos}}

\code{\link{get_adduct_names}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
