% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adducts.R
\name{adductCharge}
\alias{adductCharge}
\title{Get the charge of an adduct}
\usage{
adductCharge(x)
}
\arguments{
\item{x}{\code{character} with the adduct definition}
}
\value{
\code{integer} of length equal to \code{x} with the charge of the adduct/ion.
}
\description{
\code{adductCharge()} returns the charge of an adduct.
}
\examples{

a <- c("[M+3H]3+", "[M+H]+", "[M-2H]2-", "[M-H]-")
adductCharge(a)
}
\seealso{
Other adduct related functions: 
\code{\link{adductFormula}()},
\code{\link{adductNames}()},
\code{\link{formula2mz}()},
\code{\link{mass2mz}()},
\code{\link{mz2mass}()},
\code{\link{standardizeSingleCharge}()}
}
\concept{adduct related functions}
