% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adducts.R
\name{adductNames}
\alias{adductNames}
\title{Retrieve names of adducts}
\usage{
adductNames(polarity = c("positive", "negative"))
}
\arguments{
\item{polarity}{\code{character} Definition of ion mode, either "positive" or "negative"}
}
\value{
\code{character} A vector of all valid adduct names for the selected ion
mode
}
\description{
\code{adductNames} allows to retrieve all valid adduct names currently supported
}
\author{
Michael Witting
}
