% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_adduct_calc.R
\name{calc_adduct_mass}
\alias{calc_adduct_mass}
\title{Calculate adduct mass from neutral exact mass

This function calculates the m/z from a given exact mass and a valid adduct defintion, e.g. [M+H]+}
\usage{
calc_adduct_mass(exact_mass, adduct)
}
\arguments{
\item{exact_mass}{neutral exact mass}

\item{adduct}{adduct definition, e.g. [M+H]+}
}
\description{
Calculate adduct mass from neutral exact mass

This function calculates the m/z from a given exact mass and a valid adduct defintion, e.g. [M+H]+
}
\examples{
library(metabolomicsUtils)
calc_adduct_mass(731.5465, "[M+H]+")

}
\seealso{
\code{\link{calc_neutral_mass}}

\code{\link{create_ion_formula}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
