% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_adduct_calc_user.R
\name{calc_adduct_mass_user}
\alias{calc_adduct_mass_user}
\title{Calculate adduct mass from neutral exact mass

This function calculates the m/z from a given exact mass and a adduct definition
defined by the user.}
\usage{
calc_adduct_mass_user(exact_mass, adduct_definition)
}
\arguments{
\item{exact_mass}{neutral exact mass}

\item{adduct_definition}{}
}
\description{
Calculate adduct mass from neutral exact mass

This function calculates the m/z from a given exact mass and a adduct definition
defined by the user.
}
\examples{
library(metabolomicsUtils)

adduct_defintion <- list("[M+H]+" = list(mass_multi = 1,
mass_add =  1.007276,
formula_add = "H",
formula_sub = "C0",
charge = 1))#' 

calc_adduct_mass(731.5465, adduct_definition)

}
\seealso{
\code{\link{calc_neutral_mass_user}}

\code{\link{create_ion_formula_user}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
