% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_adduct_calc_user.R
\name{calc_neutral_mass_user}
\alias{calc_neutral_mass_user}
\title{Calculate neutral exact mass from adduct m/z

This function calculates neutral exact mass from a given m/z and a valid adduct defintion, e.g. [M+H]+}
\usage{
calc_neutral_mass_user(ion_mass, adduct_definition)
}
\arguments{
\item{adduct_definition}{}

\item{exact_mass}{neutral exact mass}
}
\description{
Calculate neutral exact mass from adduct m/z

This function calculates neutral exact mass from a given m/z and a valid adduct defintion, e.g. [M+H]+
}
\examples{
library(metabolomicsUtils)
calc_neutral_mass(732.5538, "[M+H]+")

}
\seealso{
\code{\link{calc_adduct_mass_user}}

\code{\link{create_ion_formula_user}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
