% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendrickMass.R
\name{calculateKendrickMass}
\alias{calculateKendrickMass}
\title{Calculate Kendrick masses from exact masses}
\usage{
calculateKendrickMass(x)
}
\arguments{
\item{x}{\code{numeric} exact masses}
}
\value{
\code{numeric} calculated Kendrick masses
}
\description{
\code{calculateKendrickMass} Calculates the Kendrick mass from exact masses.
}
\examples{

calculateKendrickMass(760.5851)

}
\author{
Michael Witting
}
