% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendrickMass.R
\name{calculateKendrickMassDefect}
\alias{calculateKendrickMassDefect}
\title{Calculate Kendrick mass defects from exact masses}
\usage{
calculateKendrickMassDefect(x)
}
\arguments{
\item{x}{\code{numeric} exact masses}
}
\value{
\code{numeric} calculated kendrick mass defects
}
\description{
\code{calculateKendrickMassDefect} Calculates the Kendrick mass defects from exact masses.
}
\examples{

calculateKendrickMassDefect(760.5851)

}
\author{
Michael Witting
}
