% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemFormula.R
\name{calculateMass}
\alias{calculateMass}
\title{Calculate exact mass}
\usage{
calculateMass(x)
}
\arguments{
\item{x}{\code{character} or \code{numeric } single character or named numeric}
}
\value{
\code{numeric} Resulting exact mass
}
\description{
\code{calcExactMass} calculates the exact mass from a formula.
}
\examples{

calculateMass("C6H12O6")
calculateMass("NH3")
}
\author{
Michael Witting
}
