% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendrickMass.R
\name{calculateReferencedKendrickMassDefect}
\alias{calculateReferencedKendrickMassDefect}
\title{Calculate referenced Kendrick mass defect from exact mass}
\usage{
calculateReferencedKendrickMassDefect(x, rkmd = 0.749206)
}
\arguments{
\item{x}{\code{numeric} exact masses}

\item{rkmd}{\code{numeric}Reference Kendrick mass defect to be used for calculation}
}
\value{
\code{numeric} calculated kendrick mass defects
}
\description{
\code{calculateKendrickMassDefect} Calculates the referenced Kendrick mass defects
from exact masses.
}
\examples{

calculateReferenceKendrickMassDefect(760.5851, rkmd = 0.749206)

}
\author{
Michael Witting
}
