% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemFormula.R
\name{containsElements}
\alias{containsElements}
\title{Check if one formula is contained in another}
\usage{
containsElements(x, y)
}
\arguments{
\item{x}{\code{character} Single string with a chemical formula}

\item{y}{\code{character} Single string with a chemical formula that shall be
contained in \code{x}}
}
\value{
\code{logical} TRUE if \code{y} is contained in \code{x}
}
\description{
\code{containsElements} checks if one sum formula is contained in another.
}
\examples{
containsElements("C6H12O6", "H2O")
containsElements("C6H12O6", "NH3")
}
\author{
Michael Witting
}
