% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retentionIndexing.R
\name{correctRindex}
\alias{correctRindex}
\title{2-point correction of RIs}
\usage{
correctRindex(x, y)
}
\arguments{
\item{x}{\code{numeric} vector with retention indices}

\item{y}{\code{data.frame}data.frame containing two columns, where the first
holds the measured RIs of the reference substances and the second the
reference RIs}
}
\value{
\code{numeric} vector of same length as x with correction retention
indices. Values are floating point decimals. If integer values shall be
used conversion has to be performed manually
}
\description{
\code{correctRindex} performs correction of RIs based on reference substances
}
\examples{

ref <- data.frame(rindex = c(110, 210),
refindex = c(100, 200))
rindex <- c(110, 210)
correctRindex(rindex, ref)
}
\author{
Michael Witting
}
