% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemFormula.R
\name{countElements}
\alias{countElements}
\title{Count elements in a chemical formula}
\usage{
countElements(x)
}
\arguments{
\item{x}{\code{character(1)} representing a chemical formula.}
}
\value{
\code{integer} with the element counts (names being elements).
}
\description{
\code{countElements} parses a string representing a chemical formula into a named
vector of element counts.
}
\examples{

countElements("C6H12O6")
countElements("C11H12N2O2")
}
\seealso{
\code{\link[=pasteElements]{pasteElements()}}
}
\author{
Michael Witting
}
