% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_adduct_calc.R
\name{create_ion_formula}
\alias{create_ion_formula}
\title{Create ion formula from neutral formula and adduct

This function creates an ion formula from a neutral chemical formula and a valid adduct defintion, e.g. [M+H]+}
\usage{
create_ion_formula(chem_formula, adduct)
}
\arguments{
\item{adduct}{adduct definition, e.g. [M+H]+}

\item{exact_mass}{neutral exact mass}
}
\description{
Create ion formula from neutral formula and adduct

This function creates an ion formula from a neutral chemical formula and a valid adduct defintion, e.g. [M+H]+
}
\examples{
library(metabolomicsUtils)
create_ion_formula("C6H12O6", "[M+H]+")

}
\seealso{
\code{\link{calc_adduct_mass}}

\code{\link{calc_neutral_mass}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
