% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_adduct_calc_user.R
\name{create_ion_formula_user}
\alias{create_ion_formula_user}
\title{Create ion formula from neutral formula and adduct

This function creates an ion formula from a neutral chemical formula and a valid adduct defintion, e.g. [M+H]+}
\usage{
create_ion_formula_user(chem_formula, adduct_definition)
}
\arguments{
\item{adduct_definition}{}

\item{exact_mass}{neutral exact mass}
}
\description{
Create ion formula from neutral formula and adduct

This function creates an ion formula from a neutral chemical formula and a valid adduct defintion, e.g. [M+H]+
}
\examples{
library(metabolomicsUtils)
create_ion_formula("C6H12O6", "[M+H]+")

}
\seealso{
\code{\link{calc_adduct_mass_user}}

\code{\link{calc_neutral_mass_user}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
