% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-isotope.R
\name{.is_isotope_intensity_range}
\alias{.is_isotope_intensity_range}
\title{Title Checking the intensity}
\usage{
.is_isotope_intensity_range(x, m, intensity, substDefinition)
}
\arguments{
\item{x}{intensity of the matching peaks. x has length equal to the number
of rows of substDefinition. The i-th element of x represent the intensity
associated to the peak whose m/z difference is associated to the i-th mzd in
substDefinition if any or NA.}

\item{m}{mass of the current (assumed monoisotopic) peak}

\item{intensity}{intensity of the current (assumed monoisotopic) peak}

\item{substDefinition}{substitutions definition data.frame}
}
\value{
indexes of the intensities in x that are part of a isotopic group
}
\description{
Title Checking the intensity
}
