% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-isotope.R
\name{.isotope_peaks}
\alias{.isotope_peaks}
\title{Identfying isotopologue peaks in MS data}
\usage{
.isotope_peaks(
  x,
  substDefinition = isotopicSubstitutionMatrix(),
  tolerance = 0,
  ppm = 20,
  seedMz = numeric(),
  charge = 1
)
}
\arguments{
\item{x}{\code{matrix} with spectrum data (columns \code{mz} and \code{intensity}).}

\item{substDefinition}{\code{matrix} with isotopes definition (columns
\code{name}, \code{degree}, \code{md}, \code{min_slope}, \code{max_slope}). This matrix
has to have rows ordered in such a way that the column \code{md} is sorted in
increasing order.}

\item{tolerance}{numeric representing the tolerance for the relaxed matching
of m/z values of peaks}

\item{ppm}{numeric(1) representing a relative, value-specific
parts-per-million (PPM) tolerance for the relaxed matching of m/z values of
peaks}

\item{seedMz}{numeric ordered vector containing m/z values. If provided,
the function checks if there are peaks in x whose m/z match them. If so,
it looks for isotope groups related to this subset of peaks.}

\item{charge}{numeric(1) representing the charge of the ionized compounds}
}
\value{
list of vectors. Each vector in the returned list contains the
indexes of the rows in \code{x} that match a certain isotope group found by
the function.
}
\description{
Given a spectrum (i.e. a peak matrix with m/z and intensity values)
the function identifies groups of peaks that correspond to isotopes
}
\details{
The function iterates over the peaks (rows) in x. Firstly, it checks
the presence of peaks whose m/z difference with respect to the current peak
matches (the matching can be relaxed via the parameters \code{ppm} and
\code{tolerance}) certain m/z differences related to mass differences of
certain substitutions in \code{substDefinition} and the chosen \code{charge}. Then, if
any such peak is found, the function checks if also their intensity is
compatible with them being part of a isotopic group and if so they are
grouped together. When some peaks are grouped together their indexes are
excluded from the set of indexes that are searched for further groups.
}
\examples{

}
