% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{fit_lm}
\alias{fit_lm}
\title{Fit a linear model to each row of a provided data matrix}
\usage{
fit_lm(
  formula,
  data,
  y,
  method = c("lm", "lmrob"),
  control = NULL,
  minVals = ceiling(nrow(data) * 0.75),
  model = TRUE,
  ...,
  BPPARAM = BiocParallel::SerialParam()
)
}
\value{
A \code{list} with linear models (either of type \emph{lm} or \emph{lmrob}) or
length equal to the number of rows of \code{y}. \code{NA} is reported for rows
with too few non-missing data points (depending on parameter
\code{minValues}).
}
\description{
\code{fit_lm} allows to fit a linear regression model (defined with parameter
\code{formula}) to each row of the numeric data matrix submitted with parameter
\code{y}. Additional covariates used in each model are expected to be provided
as a \code{data.frame} with parameter \code{data}.

The linear model is expected to be defined by a formula starting with
\verb{y ~ }. To model for example an injection index dependency of values in
\code{y} a formula \code{y ~ injection_index} could be used, with values for the
injection index being provided as a column \code{"injection_index"} in the
\code{data} data frame. \code{fit_lm} would thus fit this model to each row of
\code{y}.

Linear models can be fitted either with the standard least squares of
\code{\link[=lm]{lm()}} by setting \code{method = "lm"} (the default), or with the more robust
methods from the \emph{robustbase} package with \code{method = "lmrob"}.
}
\author{
Johannes Rainer
}
