% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_formula_math.R
\name{formula_addition}
\alias{formula_addition}
\title{add two chemical formula

This function add one formula to another.}
\usage{
formula_addition(target_chem_formula, query_chem_formula)
}
\arguments{
\item{target_chem_formula}{Single string with chemical formula}

\item{query_chem_formula}{Single string with chemical formula}
}
\description{
add two chemical formula

This function add one formula to another.
}
\examples{
library(metabolomicsUtils)
formula_addition("C6H12O6", "H2O")

}
\seealso{
\code{\link{contains_formula}}

\code{\link{formula_subtraction}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
