% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_formula_math.R
\name{formula_subtraction}
\alias{formula_subtraction}
\title{subtract two chemical formula

This function subtracts one formula from another.}
\usage{
formula_subtraction(target_chem_formula, query_chem_formula)
}
\arguments{
\item{target_chem_formula}{Single string with chemical formula}

\item{query_chem_formula}{Single string with chemical formula that should be subtracted from the targetFormula}
}
\description{
subtract two chemical formula

This function subtracts one formula from another.
}
\examples{
library(metabolomicsUtils)
formula_subtraction("C6H12O6", "H2O")

}
\seealso{
\code{\link{contains_formula}}

\code{\link{formula_addition}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
