% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_adduct_rules.R
\name{get_adduct_names}
\alias{get_adduct_names}
\title{Return current adduct naming

This function returns either all or only positive or negative ionization 
mode adduct names used in this package.}
\usage{
get_adduct_names(mode = "all")
}
\arguments{
\item{mode}{Ionization mode for which the adduct names shall be returned, 
either \code{all}, \code{positive} or \code{negative}}
}
\description{
Return current adduct naming

This function returns either all or only positive or negative ionization 
mode adduct names used in this package.
}
\seealso{
\code{\link{adduct_rules}}

\code{\link{adduct_rules_pos}}

\code{\link{adduct_rules_neg}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
