% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_adduct_synonyms.R
\name{get_adduct_synonym}
\alias{get_adduct_synonym}
\title{TODO user defined libraries?}
\usage{
get_adduct_synonym(adduct, adduct_library = "Agilent", direction = "from")
}
\description{
TODO user defined libraries?
}
