% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRT.R
\name{getRtime}
\alias{getRtime}
\title{Get retention time value of highest intensity}
\usage{
getRtime(x, minInt)
}
\arguments{
\item{x}{\code{data.frame}data.frame containing two columns, where the first
holds the retention times and the second the corresponding intensities}

\item{minInt}{\code{numeric} a single numeric that defines the minimum intensity
a peak must have. If no peak is present in that RT range, or if its intensity
falls below the given threshold, an error is returned.}
}
\value{
\code{numeric} a single numeric is returned, which holds the retention time
at the maximum of given intensity.
}
\description{
\code{getRtime} uses retention times (RTs) and their intensities to determine the
RT of the highest intensity within that RT range and below a defined
intensity threshold. The intensity threshold ensures that a real peak is
being picked and not any noise if no peak is present in that RT range.
}
\examples{

rti <- data.frame(rtime = c(10,20,30,40,50,60,70,80,90,100),
                 intensity = c(50,50,100,180,200,200,90,50,50,50))

getRtime(rti, minInt = 60)

}
\author{
Liesa Salzer
}
