% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{guessSource}
\alias{guessSource}
\title{Guess the source of column names}
\usage{
guessSource(x = character())
}
\arguments{
\item{x}{A \code{character} vector of column names for which the source should be
guessed.}
}
\value{
A \code{character(1)} with the name of the guessed source software.
}
\description{
Guess the source of column names based on predefined mappings. The function
counts how many elements in the input vector \code{x} match the column names
defined in the mapping schema for each software and returns the software with
the highest match count.
}
\examples{

## MS-Dial names
x <- c("Average Rt(min)", "Alignment ID", "Average Mz")
guessSource(x)

}
\author{
Gabriele Tomè
}
