% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalStandards.R
\name{internalStandardMixNames}
\alias{internalStandardMixNames}
\title{Get names of internal standard mixes supported}
\usage{
internalStandardMixNames()
}
\value{
\code{character} names of available IS mixes
}
\description{
\code{mixNames} returns a table with metabolite standards available
in commercial internal standard mixes
}
\examples{

mixNames()
}
\author{
Michael Witting
}
