% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-isotope.R
\name{isotopologues}
\alias{isotopologues}
\title{Identfying isotopologue peaks in MS data}
\usage{
isotopologues(
  x,
  substDefinition = isotopicSubstitutionMatrix(),
  tolerance = 0,
  ppm = 20,
  seedMz = numeric(),
  charge = 1
)
}
\arguments{
\item{x}{\code{matrix} with spectrum data (columns \code{mz} and \code{intensity}).}

\item{substDefinition}{\code{matrix} or \code{data.frame} with definition of isotopic
substitutions (columns \code{"name"}, \code{"degree"}, \code{"md"}, \code{"min_slope"},
\code{"max_slope"}). The rows in this matrix have to be ordered by column
\code{md} in increasing order. See \code{\link[=isotopicSubstitutionMatrix]{isotopicSubstitutionMatrix()}} for more
information on the format and content.}

\item{tolerance}{\code{numeric(1)} representing the absolute tolerance for the
relaxed matching of m/z values of peaks. See \code{\link[MsCoreUtils:matching]{MsCoreUtils::closest()}} for
details.}

\item{ppm}{\code{numeric(1)} representing a relative, value-specific
parts-per-million (PPM) tolerance for the relaxed matching of m/z values
of peaks. See \code{\link[MsCoreUtils:matching]{MsCoreUtils::closest()}} for details.}

\item{seedMz}{\code{numeric} vector of \strong{ordered} m/z values. If provided,
the function checks if there are peaks in \code{x} which m/z match them.
Isotope groups are then defined only for this subset of peaks.}

\item{charge}{\code{numeric(1)} representing the expected charge of the ionized
compounds.}
}
\value{
\code{list} of \code{integer} vectors. Each \code{integer} vector contains the
indixes of the rows in \code{x} with potential isotopologues of the same
compound.
}
\description{
Given a spectrum (i.e. a peak matrix with m/z and intensity values)
the function identifies groups potential isotopologue peaks based on
pre-defined mass differences and intensity (probability) ratios. These
parameters defining the isotopic substitutions characterizing the
isotopologues need to be passed to the function with parameter
\code{substDefinition}.
}
\details{
The function iterates over the peaks (rows) in \code{x}. For each peak (which is
assumed to be the monoisotopic peak) it searches other peaks in \code{x} with a
difference in m/z matching (given \code{ppm} and \code{tolerance}) any of the
pre-defined mass differences in \code{substDefinitions} (column \code{"md"}).

For matching peaks, the function next evaluates whether the intensity is
within the expected (pre-defined) intensity range. Using \code{"min_slope"} and
\code{"max_slope"} for the respective potentially matching isotopic substitution
in \code{substDefinition}, the function estimates a (m/z dependent) lower and
upper intensity ratio limit based on the peak's m/z.

Peaks already assigned to an itotope group are not considered in  from
difference  if also their intensity is
compatible with them being part of a isotopic group and if so they are
grouped together. When some peaks are grouped together their indexes are
excluded from the set of indexes that are searched for further groups.
}
\author{
Andrea Vicini
}
