% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendrickMass.R
\name{isRkmd}
\alias{isRkmd}
\title{Check if referenced Kendrick mass defect falls within a specific error
range}
\usage{
isRkmd(x, rkmdTolerance = 0.1)
}
\arguments{
\item{x}{\code{numeric} referenced Kendrick mass defects as calculated by
\code{calculateReferencedKendrickMassDefect()}}

\item{rkmdTolerance}{\code{numeric} tolerance for chekcking if reference Kendrick
mass defect falls within a specific range}
}
\value{
\code{boolean} indicating if RKMD falls within range
}
\description{
\code{calculateKendrickMassDefect} Calculates the Kendrick mass defects from exact masses.
}
\examples{

calculateKendrickMassDefect(760.5851)

}
\author{
Michael Witting
}
