% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{mappingSchema}
\alias{mappingSchema}
\title{Get mapping schema for column name translation}
\usage{
mappingSchema()
}
\value{
A \code{data.frame} with the mapping schema.
}
\description{
Get the mapping schema as a \code{data.frame} that defines the mapping between
column names of different software.
}
\examples{

mappingSchema()

}
\author{
Gabriele Tomè
}
