% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{matchSimilar}
\alias{matchSimilar}
\title{Match numeric values allowing an acceptable difference}
\usage{
matchSimilar(x, y, tolerance = 0, ppm = 0)
}
\arguments{
\item{x}{\code{numeric} with values to match against \code{y}.}

\item{y}{\code{numeric} with values to match elements in \code{x} against.}

\item{tolerance}{\code{numeric} with an absolute acceptable difference of values
to be considered matching. Has to be either a \code{numeric} of length 1 or
length equal \code{length(x)} (to specify a different tolerance for each
input value). Defaults to \code{tolerance = 0}.}

\item{ppm}{\code{numeric(1)} defining a relative acceptable difference (in parts
per million) for values to be considered matching. This is applied to
values in \code{x}. Defaults to \code{ppm = 0}.}
}
\value{
\code{list} (length equal length of \code{x}) with \code{integer} values. Each
element in this returned \code{list} representing the index of the values in
\code{y} that match the respective value in \code{x} given the acceptable
differences defined by \code{tolerance} and \code{ppm}. For values in \code{x} that
do not match any value in \code{y} \code{integer()} (integer of length 0) is
returned.
}
\description{
\code{matchSimilar} allows to match values in \code{x} against values in \code{y} allowing
differences between values as specified with parameters \code{tolerance} and
\code{ppm}. Note that, in contrast to the \code{\link[base:match]{base::match()}} function, indices of
\strong{all} matches of each element in \code{x} in \code{y} is returned.
}
\examples{

yvals <- abs(rnorm(1000))
yvals[123] <- yvals[2] - yvals[2] * 10 * 1e-6
yvals[124] <- yvals[2] + yvals[2] * 10 * 1e-6
yvals[125] <- yvals[2] + yvals[2] * 12 * 1e-6
xvals <- yvals[c(2, 3, 3, 20, 21, 20)]
xvals[2] <- xvals[2] + (10 * xvals[2] / 1e6)
xvals[3] <- xvals[3] - (10 * xvals[3] / 1e6)
xvals[6] <- xvals[6] + (12 * xvals[6] / 1e6)

## Perfect matches:
matchSimilar(xvals, yvals)

## Match allowing +/- 10ppm difference
matchSimilar(xvals, yvals, ppm = 10)

## Match allowing +/- 20ppm difference
matchSimilar(xvals, yvals, ppm = 20)

## perfect matches only
matchSimilar(1:3, c(1.1, 2, 3.1, 2.1, 1))

## matches with a constant tolerance
matchSimilar(1:3, c(1.1, 2, 3.1, 2.1, 1), tolerance = 0.1)

## matches with a different tolerance for each
matchSimilar(1:3, c(1.1, 2, 3.1, 2.1, 1), tolerance = c(0.1, 0, 0.1))
}
\author{
Johannes Rainer
}
