% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclosest.R
\name{mclosest}
\alias{mclosest}
\title{Extract closest values in a pairwise manner between two matrices}
\usage{
mclosest(x, table, ppm = 0, tolerance = Inf)
}
\arguments{
\item{x}{\code{numeric}  matrix or data frame representing the query data. each
row will be compared to 'table'. Need to have the same numbers of rows and in
the same order as 'table'.}

\item{table}{\code{numeric} matrix or data frame containing the reference data to
be matched with each row of 'x'. Need to have the same numbers of rows and in
the same order as 'x'.}

\item{ppm}{\code{numeric} representing a relative, value-specific
parts-per-million (PPM) tolerance that is added to tolerance (default is 0).}

\item{tolerance}{\code{numeric} accepted tolerance.}
}
\value{
\code{numeric} vector of indices indicating the closest row of 'table' for
each row of 'x'.
}
\description{
\code{mclosest} functions calculate the closest rows in a pairwise manner.
It returns the index of the closest row in 'table' for each row in 'x'.
}
\details{
If, for a row of 'x', 2 rows of 'table' are closest only the index of first
row will be returned.

For both the 'tolerance' and 'ppm' arguments, if their length is different to
the number of rows of 'x' and 'table', the input argument will be replicated
to match it.
}
\examples{
x <- data.frame(a = 1:5, b = 3:7)
table <- data.frame(c = c(11, 23, 3, 5, 1), d = c(32:35, 45))
ppm <- 0.5
tolerance <- 25

}
\author{
Philippine Louail
}
