% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemFormula.R
\name{multiplyElements}
\alias{multiplyElements}
\title{Multiply chemical formulas by a scalar}
\usage{
multiplyElements(x, k)
}
\arguments{
\item{x}{\code{character} strings with chemical formula}

\item{k}{\code{numeric(1)} positive integer by which each formula will be
multiplied.}
}
\value{
\code{character} strings with the standardized chemical formula.
}
\description{
\code{multiplyElements} Multiply the number of atoms of each element by a
constant, positive, integer
}
\examples{

multiplyElements("H2O", 3)

multiplyElements(c("C6H12O6", "Na", "CH4O"), 2)

}
\author{
Roger Gine
}
