% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adducts.R
\name{mz2mass}
\alias{mz2mass}
\title{Calculate neutral mass}
\usage{
mz2mass(x, adduct = "[M+H]+")
}
\arguments{
\item{x}{\code{numeric} m/z value for which the neutral mass shall be calculated.}

\item{adduct}{\code{character} specifying the name of the adduct;
supported values are returned by \code{\link[=adductNames]{adductNames()}}}
}
\value{
\code{numeric} representing the calculated neutral mass.
}
\description{
\code{mz2mass} calculates the neutral mass from a given m/z value and adduct
definition
}
\examples{

ion_mass <- 100
adduct <- "[M+H]+"

## Calculate m/z of [M+H]+ adduct from neutral mass
mz2mass(ion_mass, adduct)

ion_mass <- 100
adduct <- "[M+Na]+"

## Calculate m/z of [M+Na]+ adduct from neutral mass
mz2mass(ion_mass, adduct)
}
\seealso{
\code{\link[=mass2mz]{mass2mz()}} for the reverse calculation, \code{\link[=adductNames]{adductNames()}} for
supported adduct definitions.
}
\author{
Michael Witting
}
