% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_massdiff_rules.R
\name{read_massdiff_table}
\alias{read_massdiff_table}
\title{Read mass differences from file}
\usage{
read_massdiff_table(file, ...)
}
\description{
Read mass differences from file
}
\details{
create a list from a user-input file that is in the same format as 
massdiff_list (see above)
}
