% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{softwareMapping}
\alias{softwareMapping}
\title{Get supported software for column name translation}
\usage{
softwareMapping()
}
\value{
A \code{character} vector with the names of the supported software.
}
\description{
Get the names of the supported software defined in the mapping schema
}
\examples{

softwareMapping()

}
\author{
Gabriele Tomè
}
