% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{softwareMappingSchema}
\alias{softwareMappingSchema}
\title{Get mapping schema for name translation}
\usage{
softwareMappingSchema(path = NULL)
}
\arguments{
\item{path}{Optional \code{character(1)} with the path to a custom mapping schema
file in TSV format. If not provided, the default mapping schema defined
in the package will be used. The file must have a header row with
software names as names and the corresponding names as values.}
}
\value{
A \code{data.frame} with the mapping schema.
}
\description{
Get the mapping schema as a \code{data.frame} that defines the mapping between
names of different software.
}
\examples{

softwareMappingSchema()

}
\seealso{
Other name translation functions: 
\code{\link{guessSource}()},
\code{\link{nameMapping}()},
\code{\link{softwareMapping}()},
\code{\link{translate}()}
}
\author{
Gabriele Tomè
}
\concept{name translation functions}
