% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_formula_parser.R
\name{standardize_formula}
\alias{standardize_formula}
\title{Create chemical formula from parsed list

This function standardizes a supplied chemical chem_formula according to the Hill notation system.}
\usage{
standardize_formula(chem_formula)
}
\arguments{
\item{chem_formula}{Single string with chemical chem_formula}
}
\description{
Create chemical formula from parsed list

This function standardizes a supplied chemical chem_formula according to the Hill notation system.
}
\examples{
library(metabolomicsUtils)
standardizeChemFormula("O6C6H12")

}
\seealso{
\code{\link{formula_to_list}}

\code{\link{list_to_formula}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
