% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemFormula.R
\name{standardizeFormula}
\alias{standardizeFormula}
\title{Create chemical formula from a named vector}
\usage{
standardizeFormula(x)
}
\arguments{
\item{x}{\code{character} Single string with the chemical formula to standardize}
}
\value{
\code{character} Single string with the standardized chemical formula
}
\description{
\code{standardizeFormula} standardizes a supplied chemical chem_formula according to the Hill notation system.
}
\examples{

standardizeFormula("C6O6H12")
}
\author{
Michael Witting
}
