% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adducts.R
\name{standardizeSingleCharge}
\alias{standardizeSingleCharge}
\title{Fix short-hand charge notation in adduct definition}
\usage{
standardizeSingleCharge(x)
}
\arguments{
\item{x}{\code{character} with adduct definitions}
}
\value{
\code{character} with standardized single charge definitions.
}
\description{
\code{standardizeSingleCharge()} replaces the short-hand notation of single
charges eventually present in adduct definition with the \emph{standard}
notion, i.e., it replaces \code{"+"} with \code{"1+"} in e.g. \code{"[M+H]+"} and \code{"-"}
with \code{"1-"} in \code{"[M-H]-"}.
}
\examples{

a <- c("[M+H]+", "[M-H]-", "[M+H]1+")
standardizeSingleCharge(a)
}
\seealso{
Other adduct related functions: 
\code{\link{adductCharge}()},
\code{\link{adductFormula}()},
\code{\link{adductNames}()},
\code{\link{formula2mz}()},
\code{\link{mass2mz}()},
\code{\link{mz2mass}()}
}
\concept{adduct related functions}
