% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemFormula.R
\name{subtractElements}
\alias{subtractElements}
\title{subtract two chemical formula}
\usage{
subtractElements(x, y)
}
\arguments{
\item{x}{\code{character} Single string with chemical formula}

\item{y}{\code{character}  Single or multiple strings with chemical formula that
should be subtracted from \code{x}}
}
\value{
\code{character} Resulting formula
}
\description{
\code{subtractElements} subtracts one chemical formula from another.
}
\examples{
subtractElements("C6H12O6", "H2O")
subtractElements("C6H12O6", "NH3")
}
\author{
Michael Witting
}
