% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemFormula.R
\name{addElements}
\alias{addElements}
\title{Combine chemical formulae}
\usage{
addElements(x, y = NA_character_)
}
\arguments{
\item{x}{\code{character} Vector with 1 or more chemical formulae to be added}

\item{y}{\code{character} Vector with 1 or more chemical formulae to be added}
}
\value{
\code{character} Resulting formula
}
\description{
\code{addElements} Add one chemical formula to another.
}
\examples{

addElements("C6H12O6", "Na")

addElements("C6H12O6", c("Na", "H2O"))
}
\author{
Michael Witting
}
