% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemFormula.R
\name{addElements}
\alias{addElements}
\title{Combine chemical formulae}
\usage{
addElements(x, y)
}
\arguments{
\item{x}{\code{character} strings with chemical formula}

\item{y}{\code{character} strings with chemical formula that
should be added from \code{x}}
}
\value{
\code{character} Resulting formula
}
\description{
\code{addElements} Add one chemical formula to another.
}
\examples{

addElements("C6H12O6", "Na")

addElements("C6H12O6", c("Na", "H2O"))
}
\author{
Michael Witting and Sebastian Gibb
}
