% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemFormula.R
\name{calculateMass}
\alias{calculateMass}
\title{Calculate exact mass}
\usage{
calculateMass(x)
}
\arguments{
\item{x}{\code{character} representing chemical formula(s) or a \code{list } of
\code{numeric} with element counts such as returned by \code{\link[=countElements]{countElements()}}.}
}
\value{
\code{numeric} Resulting exact mass.
}
\description{
\code{calculateMass} calculates the exact mass from a formula.
}
\examples{

calculateMass("C6H12O6")
calculateMass("NH3")
calculateMass(c("C6H12O6", "NH3"))
calculateMass(c("C6H12O6", "[13C3]C3H12O6"))
}
\author{
Michael Witting
}
