% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemFormula.R
\name{countElements}
\alias{countElements}
\title{Parse a chemical formula into a named vector}
\usage{
countElements(x)
}
\arguments{
\item{x}{\code{character} Single string with chemical}
}
\value{
\code{numeric} A named list containing the parsed chemical formula
}
\description{
\code{countElements} This function parses a chemical into a named vector
}
\examples{

countElements("C6H12O6")
countElements("C11H12N2O2")
}
\author{
Michael Witting
}
