% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalStandards.R
\name{internalStandardMixNames}
\alias{internalStandardMixNames}
\title{Get names of internal standard mixes provided by the package}
\usage{
internalStandardMixNames()
}
\value{
\code{character} names of available IS mixes
}
\description{
\code{internalStandardMixNames} returns available names of internal standard
mixes provided by the \code{MetaboCoreUtils} package.
}
\examples{

internalStandardMixNames()
}
\author{
Michael Witting
}
