% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalStandards.R
\name{internalStandards}
\alias{internalStandards}
\title{Get internal standards}
\usage{
internalStandards(mix = "QReSS")
}
\arguments{
\item{mix}{\code{character} Name of the internal standard mix that shall be returned}
}
\value{
\code{data.frame} data on internal standards
}
\description{
\code{internalStandards} returns a table with metabolite standards available
in commercial internal standard mixes
}
\examples{

internalStandards(mix = "QReSS")
internalStandards(mix = "UltimateSplashOne")
}
\author{
Michael Witting
}
