% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-isotope.R
\name{isotopicSubstitutionMatrix}
\alias{isotopicSubstitutionMatrix}
\title{Title HMBD isotope substitution}
\usage{
isotopicSubstitutionMatrix()
}
\value{
\code{data.frame} with columns \code{mzd}, \code{min_ratio} and \code{max_ratio}
}
\description{
Loads the HMBD isotope substitution
}
