% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{nameMapping}
\alias{nameMapping}
\title{Get mapping vector between two software}
\usage{
nameMapping(from = character(), to = character())
}
\arguments{
\item{from}{\code{character(1)} with the name of the source software.}

\item{to}{\code{character(1)} with the name of the target software.}
}
\value{
A named \code{character} vector with the mapping between the two software.
}
\description{
Get a named \code{character} vector that defines the mapping between column names
of two software based on the mapping schema. The names of the returned vector
are the column names of the \code{from} software and the values are the
corresponding column names of the \code{to} software.
}
\examples{

nameMapping(from = "MS-Dial", to = "mzTab-M")

}
\author{
Gabriele Tomè
}
