% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemFormula.R
\name{pasteElements}
\alias{pasteElements}
\title{Create chemical formula from a named vector}
\usage{
pasteElements(x)
}
\arguments{
\item{x}{\code{character} A named vector containing the number of elements}
}
\value{
\code{character} Single string with the chemical formula
}
\description{
\code{pasteElements} This function creates a chemical formula from a parsed list
}
\examples{

elements <- c("C" = 6, "H" = 12, "O" = 6)
pasteElements(elements)
}
\author{
Michael Witting
}
