% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{softwareMapping}
\alias{softwareMapping}
\title{Get supported software for name translation}
\usage{
softwareMapping()
}
\value{
A \code{character} vector with the names of the supported software.
}
\description{
Get the names of the supported software defined in the default mapping schema
}
\examples{

softwareMapping()

}
\seealso{
Other name translation functions: 
\code{\link{guessSource}()},
\code{\link{nameMapping}()},
\code{\link{softwareMappingSchema}()},
\code{\link{translate}()}
}
\author{
Gabriele Tomè
}
\concept{name translation functions}
