% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate names based on a provided mapping}
\usage{
translate(x = character(), mapping = NULL)
}
\arguments{
\item{x}{\code{character} vector of names to be translated.}

\item{mapping}{A named \code{character} vector that defines the mapping for
translation. The names of the vector should be the original names
and the values should be the translated names.}
}
\value{
A \code{character} vector with the translated names.
}
\description{
Map names from one software to another. The function replaces elements
in the input vector \code{x} with their corresponding values in the mapping if a
match is found. If an element in \code{x} does not have a corresponding mapping,
it is returned unchanged with a warning.
}
\examples{

## MS-Dial names
x <- c("Average Rt(min)", "Alignment ID", "Average Mz")
map_vec <- nameMapping(from = "MS-Dial", to = "mzTab-M")
translate(x, mapping = map_vec)

}
\seealso{
Other name translation functions: 
\code{\link{guessSource}()},
\code{\link{nameMapping}()},
\code{\link{softwareMapping}()},
\code{\link{softwareMappingSchema}()}
}
\author{
Gabriele Tomè
}
\concept{name translation functions}
